+++ sdk/confidentialledger/azure-security-confidentialledger/src/samples/java/com/azure/security/confidentialledger/DeleteUserSample.java
@@ -5,17 +5,19 @@ package com.azure.security.confidentialledger;
 
 import com.azure.core.http.rest.RequestOptions;
 import com.azure.core.http.rest.Response;
-import com.azure.core.util.BinaryData;
 import com.azure.identity.DefaultAzureCredentialBuilder;
 
-public class GetCurrentLedgerEntry {
+public class DeleteUserSample {
     public static void main(String[] args) {
         ConfidentialLedgerClient confidentialLedgerClient =
                 new ConfidentialLedgerClientBuilder()
                         .credential(new DefaultAzureCredentialBuilder().build())
-                        .ledgerUri("https://my-ledger.confidential-ledger.azure.com")
+                        .ledgerEndpoint("https://my-ledger.confidential-ledger.azure.com")
                         .buildClient();
         RequestOptions requestOptions = new RequestOptions();
-        Response<BinaryData> response = confidentialLedgerClient.getCurrentLedgerEntryWithResponse(requestOptions);
+        
+        // you can retrieve your object id by going to Azure Active Directory and finding your profile
+        String aadObjectId = "<YOUR AAD ID>";
+        Response<Void> response = confidentialLedgerClient.deleteUserWithResponse(aadObjectId, requestOptions);
     }
 }