+++ sdk/core/azure-core/src/test/java/com/azure/core/implementation/AsynchronousByteChannelWriteSubscriberTests.java
@@ -3,6 +3,7 @@
 
 package com.azure.core.implementation;
 
+import com.azure.core.util.IOUtils;
 import org.junit.jupiter.api.Test;
 import org.reactivestreams.Subscription;
 import reactor.core.publisher.MonoSink;
@@ -20,9 +21,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 /**
- * Tests {@link FileWriteSubscriber}.
+ * Tests {@link AsynchronousByteChannelWriteSubscriber}.
  */
-public class FileWriteSubscriberTests {
+public class AsynchronousByteChannelWriteSubscriberTests {
     @SuppressWarnings("unchecked")
     @Test
     public void multipleSubscriptionsCancelsLaterSubscriptions() {
@@ -40,7 +41,8 @@ public class FileWriteSubscriberTests {
 
         MonoSink<Void> sink = (MonoSink<Void>) mock(MonoSink.class);
 
-        FileWriteSubscriber fileWriteSubscriber = new FileWriteSubscriber(channel, 0, sink);
+        AsynchronousByteChannelWriteSubscriber fileWriteSubscriber = new AsynchronousByteChannelWriteSubscriber(
+            IOUtils.toAsynchronousByteChannel(channel, 0), sink);
         Subscription subscription1 = mock(Subscription.class);
         Subscription subscription2 = mock(Subscription.class);